Refactor some stuff out of run_tokenizer
